Optimize the way in which the snake advances
Instead of creating a new head with newNode() and destroying the oldTail
now the snake advances by popping the tail and reappending it as the new
haed, see ouroboros(). This saves a call to malloc() and free() on each step.
newNode() is still called when growing.

Variables that hold the state of the Snake (growing, oldTail) are now part
of the object.